+++ Environment_Engine/Query/CullDuplicates.cs
@@ -40,29 +40,6 @@ namespace BH.Engine.Environment
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Removes panels which overlap each other")]
-        [Input("panels", "A collection of Environment Panels")]
-        [Output("panels", "A collection of Environment Panels with no overlaps")]
-        public static List<Panel> CullOverlaps(this List<Panel> panels)
-        {
-            List<Panel> ori = new List<Panel>(panels);
-            List<Panel> toReturn = new List<Panel>();
-            
-            while(ori.Count > 0)
-            {
-                Panel current = ori[0];
-                List<Panel> overlaps = current.IdentifyOverlaps(panels);
-
-                foreach (Panel be in overlaps)
-                    ori.Remove(be);
-
-                toReturn.Add(current);
-                ori.RemoveAt(0);
-            }
-
-            return toReturn;
-        }
-
         [Description("Removes panels which are duplicates")]
         [Input("panels", "A collection of Environment Panels")]
         [Output("panels", "A collection of Environment Panels with no duplicates")]
@@ -113,30 +90,6 @@ namespace BH.Engine.Environment
             }
 
             return panelsAsSpaces.Where(x => x.Count > 0).ToList();
-        }
-
-        [Description("Removes duplicate lines from the collection")]
-        [Input("lines", "The nested collection of lines to cull duplicates from")]
-        [Output("lines", "A collection of lines with no duplicates")]
-        public static List<Line> CullDuplicateLines(this List<Line> lines, double tolerance = Tolerance.Distance)
-        {
-            double sqTol = tolerance * tolerance;
-            List<Line> result = lines.Select(l => l).ToList();
-            for (int i = lines.Count - 2; i >= 0; i--)
-            {
-                for (int j = lines.Count - 1; j > i; j--)
-                {
-                    Line l1 = lines[i];
-                    Line l2 = lines[j];
-                    {
-                        if ((l1.Start.SquareDistance(l2.Start) <= sqTol && l1.End.SquareDistance(l2.End) <= sqTol) || (l1.Start.SquareDistance(l2.End) <= sqTol && l1.End.SquareDistance(l2.Start) <= sqTol))
-                        {
-                            lines.RemoveAt(j);
-                        }
-                    }
-                }
-            }
-            return lines;
         }
     }
 }
